nostr: adj. kind to be `u16` instead of `u64` according to NIP01
https://github.com/nostr-protocol/nips/blob/403b5199a490b6a148063003e00924f5e79ba36c/01.md#events-and-signatures

+++ bindings/nostr-ffi/src/event/raw.rs
@@ -17,7 +17,7 @@ pub struct RawEventRecord {
     /// Timestamp (seconds)
     pub created_at: u64,
     /// Kind
-    pub kind: u64,
+    pub kind: u16,
     /// Vector of strings
     pub tags: Vec<Vec<String>>,
     /// Content
