+++ crates/nostr/src/event/raw.rs
@@ -77,7 +77,7 @@ pub struct RawEvent {
     /// Timestamp (seconds)
     pub created_at: u64,
     /// Kind
-    pub kind: u64,
+    pub kind: u16,
     /// Vector of strings
     pub tags: Vec<Vec<String>>,
     /// Content
